requirements.txt
Telethon~=1.36.0
python-dotenv~=1.0.1
psycopg2-binary
sentence-transformers